Alterada funcionalidade de criar ControlDevice para criar o dispositivo para o usuário que estiver authenticado, e também alterada funcionalidade para somente apagar ControlDevice se este pertencer realmente ao usuário authenticado

+++ src/main/java/com/dougfsilva/iotnizer/service/controlDevice/DeleteControlDevice.java
@@ -2,7 +2,9 @@ package com.dougfsilva.iotnizer.service.controlDevice;
 
 import org.springframework.stereotype.Service;
 
+import com.dougfsilva.iotnizer.exception.OperationNotPermittedException;
 import com.dougfsilva.iotnizer.model.ControlDevice;
+import com.dougfsilva.iotnizer.model.User;
 import com.dougfsilva.iotnizer.mqtt.TopicManagerMqtt;
 import com.dougfsilva.iotnizer.repository.ControlDeviceRepository;
 
@@ -21,8 +23,11 @@ public class DeleteControlDevice {
 		this.topicManagerMqtt = topicManagerMqtt;
 	}
 
-	public void delete(String id) {
+	public void delete(User user, String id) {
 		ControlDevice device = findControlDevice.findById(id);
+		if(!user.getId().equals(device.getUser_id())) {
+			throw new OperationNotPermittedException("Deleting devices belonging to another user is not allowed!");
+		}
 		repository.delete(device);
 		topicManagerMqtt.removeTopic(device);
 	}
